+++ client/src/app/(withauth)/products/ledgerscan/page.tsx
@@ -18,7 +18,7 @@ import "react-json-view-lite/dist/index.css"
 export default function Page() {
   const [api, setApi] = useState("")
   const [response, setReseponse] = useState({})
-  const products = useQuery("get-products", `${endPoints.getProductConfig}?searchQuery=ledgerscan`, HTTPMethods.GET)
+  const products = useQuery(["products"], `${endPoints.getProductConfig}?searchQuery=ledgerscan`, HTTPMethods.GET)
   const selectedProduct = products?.data?.find((product: any) => product.productName === "ledgerscan")
   const [isLoading, setLoading] = useState(false)